Show full history of OOBTree objects (LP#474334).
The implementation is ugly and the tests are incomplete.

Things that deal with btrees ought to be moved into a separate module.

_loadState() should become a member of IObjectHistory.

UI bug: if btree was modified in rev #10 and one of its buckets in #12, and you
ask for rev #12 of the btree, it will show the right contents but highlight
the wrong revision in the list.