Support all kinds of ZODB databases, not just those used by Zope 3/BlueBream apps.
Fixes https://bugs.launchpad.net/zodbbrowser/+bug/494987.

The approach used is the one suggested in the bug: we create a temporary
MappingStorage for our own use, and access the database for inspection
explicitly.

We lose the ability to perform object traversal (/foo/@@zodbbrowser)
when using the standalone app mode, but the tradeoff is worth it.