Avoid IPv6 in the test suite
So listening on localhost:nnnn on a system with working IPv6 creates two
listening sockets: one for 127.0.0.1, and one for [::1].  And then
waitresss uses a MultiSocketServer, which doesn't have a singular
.effective_address, and which prints multiple 'Listening on ...' log
messages, and things break.

This happens on GitHub Actions, but not on my laptop.  Of course.
Wouldn't want the maintainer's life to lack excitement.